Function ink::codegen::execute_constructor

pub fn execute_constructor<Contract, F, R>(f: F) -> Result<(), DispatchError> where
    Contract: Storable + StorageKey + ContractEnv,
    F: FnOnce() -> R,
    <Seal<R> as ConstructorReturnType<Contract>>::Error: Encode,
    Seal<R>: ConstructorReturnType<Contract>, 

Executes the given ink! constructor.

Note

The closure is supposed to already contain all the arguments that the real constructor message requires and forwards them.