ink::codegen

Trait Env

pub trait Env {
    type EnvAccess;

    // Required method
    fn env(self) -> Self::EnvAccess;
}

Simplifies interaction with the host environment via self.

Note

This is generally implemented for storage structs that include their environment in order to allow the different dispatch functions to use it for returning the contract’s output.

Required Associated Types

type EnvAccess

The access wrapper.

Required Methods

fn env(self) -> Self::EnvAccess

Accesses the host environment with self.env() syntax.