ink::codegen

Trait TraitCallBuilder

pub trait TraitCallBuilder {
    type Builder;

    // Required methods
    fn call(&self) -> &Self::Builder;
    fn call_mut(&mut self) -> &mut Self::Builder;
}

The global call builder type for an ink! trait definition.

Required Associated Types

type Builder

The call builder type.

Required Methods

fn call(&self) -> &Self::Builder

Returns a shared reference to the global call builder type.

This allows to call &self ink! trait messages.

fn call_mut(&mut self) -> &mut Self::Builder

Returns an exclusive reference to the global call builder type.

This allows to call any ink! trait message.