ink::codegen

Trait TraitCallForwarder

pub trait TraitCallForwarder {
    type Forwarder<Abi>: TraitCallBuilder;
}

Implemented by the global trait info provider.

It is used to query the global trait call forwarder. There is one global trait call forwarder that implements the call forwarding (short- and long-form) for all calls to this trait in ink-as-dependency configuration.

Required Associated Types

type Forwarder<Abi>: TraitCallBuilder

The call forwarder type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.