ink::codegen

Module utils

Utility types and definitions used by the ink! codegen.

Macros

const_concat

Compile-time concat!-like macro that accepts const expressions as arguments (i.e. not just literals) and returns &'static str.

const_format

Compile-time format!-like macro that returns &'static str.

Structs

IsSameType

Can be used to check equality of types.

Functions

consume_type

Takes a generic type as input and just consumes it while doing nothing.