ink::codegen::utils

Function consume_type

pub const fn consume_type<T>()

Takes a generic type as input and just consumes it while doing nothing.

Note

This can be used to trigger some compile time checks due to the fact that the type consumed this way is type checked. We usually use this to make the Rust compiler check the trait bounds in particular.

Usage: Compiles

pub struct RequiresCopy<T: Copy>(PhantomData<T>);

// The following line of code works because `i32: Copy`.
consume_type::<RequiresCopy<i32>>();

Usage: Compile Error

pub struct RequiresCopy<T: Copy>(PhantomData<T>);

// The following line of code fails to compile because
// `String` does not implement `Copy`.
consume_type::<RequiresCopy<String>>();