ink
Macro selector_id

selector_id!() { /* proc-macro */ }

Computes the ink! selector of the string and expands into its u32 representation.

Note

The computation takes place at compilation time of the crate.

Example

assert_eq!(selector_id!("hello"), 843960066,);