Enum ink::reflect::DispatchError

pub enum DispatchError {
    InvalidSelector,
    UnknownSelector,
    InvalidParameters,
    CouldNotReadInput,
    PaidUnpayableMessage,
}

An error that can occur during dispatch of ink! dispatchables.

Variants

InvalidSelector

Failed to decode into a valid dispatch selector.

UnknownSelector

The decoded selector is not known to the dispatch decoder.

InvalidParameters

Failed to decode the parameters for the selected dispatchable.

CouldNotReadInput

Failed to read execution input for the dispatchable.

PaidUnpayableMessage

Invalidly paid an unpayable dispatchable.