ink::storage::traits

Trait AutoStorableHint

pub trait AutoStorableHint<Key>
where
    Key: StorageKey,
{
    type Type: Storable;
}

Automatically returns the type that should be used for storing the value.

The trait is used by codegen to determine which storage key the type should have.

Required Associated Types

type Type: Storable

Storable type with storage key inside.