Trait ink::storage::traits::Packed

pub trait Packed: Storable + Codec + Sealed { }

Trait for describing types that can be read and written to storage while all fields occupy only a single storage cell.

If at least one of the fields in the type occupies its own storage cell, this type is considered non-packed.

Note

The trait is automatically implemented for types that implement [scale::Codec] via blanket implementation.