ink::storage::traits

Trait StorableHint

pub trait StorableHint<Key>
where
    Key: StorageKey,
{
    type Type: Storable;
    type PreferredKey: StorageKey;
}

Describes the type that should be used for storing the value and preferred storage key.

Note

The trait is automatically implemented for Packed types via blanket implementation.

Required Associated Types

type Type: Storable

Storable type with storage key inside.

type PreferredKey: StorageKey

The storage key that the type prefers. It can be overwritten by an auto-generated storage key.