ink
Module codegen

Definitions and facilities mainly used by the ink! codegen.

Modules

utils

Utility types and definitions used by the ink! codegen.

Structs

DispatchInput

Used to check if T is allowed as ink! input parameter type.

DispatchOutput

Used to check if T is allowed as ink! output parameter type.

TraitMessagePayable

Used as payable property guard for ink! trait messages.

TraitMessageSelector

Used as selector property guard for ink! trait messages.

Traits

ContractCallBuilder

Used to refer to the generated contract call builder.

Env

Simplifies interaction with the host environment via self.

ImpliesReturn

Trait used as bound for the outputs of ink! trait definition messages.

StaticEnv

Simplifies interaction with the host environment via Self.

TraitCallBuilder

The global call builder type for an ink! trait definition.

TraitCallForwarder

Implemented by the global trait info provider.

TraitCallForwarderFor

Implemented by call builders of smart contracts.

TraitMessageBuilder

Access the trait message builder implementation.

Functions

deny_payment

Returns Ok if the caller did not transfer additional value to the callee.