Trait ink::codegen::ImpliesReturn

pub trait ImpliesReturn<T> { }

Trait used as bound for the outputs of ink! trait definition messages.

Note

Indicates that a value of type T or an equivalent CallBuilder instantiation of an ink! message returning a value of type T are returned by the respective ink! trait message.

The bounds are automatically generated by the #[ink::trait_definition] procedural macro.