Trait ink::codegen::StaticEnv

pub trait StaticEnv {
    type EnvAccess;

    fn env() -> Self::EnvAccess;
}

Simplifies interaction with the host environment via Self.

Note

This is generally implemented for storage structs that include their environment in order to allow the different dispatch functions to use it for returning the contract’s output.

Required Associated Types

type EnvAccess

The access wrapper.

Required Methods

fn env() -> Self::EnvAccess

Accesses the host environment with Self::env() syntax.