Trait ink::codegen::TraitCallForwarderFor

pub trait TraitCallForwarderFor<const TRAIT_ID: u32> {
    type Forwarder: TraitCallBuilder;

    // Required methods
    fn forward(&self) -> &Self::Forwarder;
    fn forward_mut(&mut self) -> &mut Self::Forwarder;
    fn build(&self) -> &<Self::Forwarder as TraitCallBuilder>::Builder;
    fn build_mut(
        &mut self
    ) -> &mut <Self::Forwarder as TraitCallBuilder>::Builder;
}

Implemented by call builders of smart contracts.

These might be implementing multiple different ink! traits. The codegen makes them implement this trait once for every ink! trait they have to implement.

While the trait is not necessary it encapsulates a lot of utility and auxiliary code required for the actual ink! trait implementations.

Required Associated Types

type Forwarder: TraitCallBuilder

Required Methods

fn forward(&self) -> &Self::Forwarder

Forwards the &self call.

Note

This is used for the short-hand calling syntax.

fn forward_mut(&mut self) -> &mut Self::Forwarder

Forwards the &mut self call.

Note

This is used for the short-hand calling syntax.

fn build(&self) -> &<Self::Forwarder as TraitCallBuilder>::Builder

Builds up the &self call.

Note

This is used for the long-hand calling syntax.

fn build_mut(&mut self) -> &mut <Self::Forwarder as TraitCallBuilder>::Builder

Builds up the &mut self call.

Note

This is used for the long-hand calling syntax.