Trait ink::Output

pub trait Output<const IS_RESULT: bool, const HANDLE_STATUS: bool, T, E>: OutputSealed {
    type ReturnType;
}

Only implemented for ValueReturned.

Used to deduce the correct return type of a chain extension method at compile time based on 2 flags: const IS_RESULT: bool and const HANDLE_STATUS: bool.

If IS_RESULT is set to false and HANDLE_STATUS is true, then type ReturnType = Result<T, E>. Otherwise type ReturnType = T.

Required Associated Types

type ReturnType